Refactored Profile Tab Panel
